Tag flow trunk based monorepo support (#1)
feat: add support for tracking tags for builds

Introduce parameter to specify a reference tag, formatted similarly to how you would enter a tag filter. ex: /^release.*/.

If a tag reference is provided the orb will now search for the previous tag that matches this reference, if one exists. It will set this as the base commit to compare the head against. If no tag, or no matching tag, is found the orb will continue as normally.

This fills a gap while using trunk based monorepos, where the trunk is used as staging and prod deployments are tagged. The original path filtering orb, even with tagged commits, only looks at the previous commit. This adds the functionality for the orb to find the previous tag matching our provided reference, and allows us to capture all commits between the current tag and the previous tag to diff against the head.

Reviewed-by: Rafe Anderson